Refactor Page dependency checking
Checking for dependencies of a Page instance is done repeatedly using
interfacing with the page instance's properties.

During a file system event when serving a MarkBind site, the site's
source files are rebuilt if the file has a "source file extension".
This unnecessarily runs the source files rebuilding process even when
such files are used as assets.

Let's extract these page dependency checking into reusable methods as
such,
and use these methods in place of the extension check.