Move trigger into a Vue component
Triggers fail to find the targeted element when it is not loaded into
the dom, when the trigger's popover and tooltip directives are being
processed.

This causes triggers to fail when the targeted element is loaded
on-demand, such as when it is inside a collapsed panel that is later
expanded.

Let's move triggers into a Vue component, allowing for the targeted
element retrieval to occur only when the trigger is activated.