Fix external dependency generation resolving
External dependencies are not regenerated if the same dependency is
being (or has finished) generated.
In such cases, the generation promise short circuits and immediately
resolves, without waiting for the same existing dependency to finish
generating.

This causes heading and keyword collection in Page to fail, which
relies on all its dependencies to have generated.
Let's return the same existing generation promise in such cases instead.