Adapt nunjucks.renderString calls to nunjucks.render
Source files are read directly by MarkBind before passing to nunjucks
for processing.
This incurs an extra unnecessary step.

Nunjucks compiles a template into a javascript function, then uses it
to render the template.
This compilation process is costly, which is alleviated by nunjucks'
internal cache.
The nunjucks.renderString call is however unable to take advantage of
this cache.

Let's utilize the nunjucks.render call instead, thereby removing the
extra step, and allowing the cache to be used.