Implement plugins system tweaks
With the merger of the include and render html stages, many breaking
changes are implied for the existing plugin interfaces.

For example, the preRender plugin hook is no longer applicable, and
postRender is no longer executed on the entire page content that is
surrounded by layout content.
Moreover, the frontMatter parameter is made inaccessible to many
plugin hooks.

Let's tweak the plugin system as such for the modified architecture.

The frontMatter parameter is removed from many plugin hooks.

In place of the plugin getSources hook - which also no longer has
access to the frontMatter, and getSpecialTags hook, let's introduce the
tagConfig plugin interface. This interfaces serves the same purpose and
consolidates the two hooks, allowing plugin authors to configure
various behaviours of tags and their attributes. This provides for an
extensible interface to implement any future tag / attribute related
plugin behaviour.

In place of the preRender hook, let's introduce the processNode plugin
hook, which taps directly into the merged html processing stage,
allowing a plugin to make modifications to any html node. This also
offers considerable performance benefits over pre / postRender, which
requires parses the entire page content.

The Plugin and PluginManager abstractions are also thus introduced, to
consolidate old and new Plugin related logic.