Merge include and render html stages
MarkBind's html processing consist two distinct stages - include and
render.

In the first, <include>s are preprocessed, then the page is rendered
with markdown-it.
This necessitates markdown-htmlparser patches, which are difficult to
write.
This also poses issues in incorporating more expressive markdown-it
plugins which heavily conflict with html parsing.

Let's combine these two html processing stages to remedy this.

The result also, is a simpler architecture surrounding a recursive
process that always follows nunjucks -> markdown -> html processing.