Implement new layout system
Layouts require the author to maintain numerous separate files (header,
footer, ...) which are combined in an opaque, restrictive manner to
form the skeleton for a page.

This increases the burden of author maintenance, with multiple files
and directories to keep track of.
It also restricts the author from forming layouts that require more
expressive structures that cannot be formed with the few fixed files.

Let's move to a simpler, one-file layout system which allows the author
to position the page content anywhere within the layout with a nunjucks
variable.
This also lifts the restrictions imposed by the way layout files are
previously combined, allowing more expressive page skeletons to be
formed.