Remove line numbers from user guide where it is redundant
Some of the code blocks in the user guide contains redundant line
numbers. There are also some file names which are not moved to the
code headers and are listed as part of the code itself.

Let's remove the redundant line numbers and format code blocks
correctly to have the file names as part of the code headers.