Rewrite panel transition for preview feature (#1430)
It is not easy to implement the panel preview feature with the existing
transition code in our panels due to the use of v-show.

Let's rewrite some of the transition code and implement our own set
of transition code to allow for more flexibility. This will allow a cleaner
implementation of the panel preview feature.