Add omitFrontmatter attribute to include (#1448)
Including a file combines the front matter of the file with the main
page. When authors wish to include the full content of the file but not
the front matter, they have to make their whole content a segment as a
work around, making the content less concise.

Let's add a new attribute to indicate whether the front matter data of
the included file should be omitted.