Remove MarkBind.afterSetup hook
The afterSetup MarkBind hook was introduced to allow user scripts and
plugin scripts to deterministically execute post Vue setup, which may
not happen synchronously, as search data was fetched before doing so.

With search data retrieval moved to post Vue setup, this is no longer
necessary.

Let's remove the hook as such.