Update deprecated Vue slot syntax (#1464)
Our internal Vue slot syntax are deprecated according to official
Vue documentation. To ensure our codebase is up-to-date, let's
convert all the deprecated syntax to the latest Vue slot shorthand
syntax that was introduced in version 2.6.0.

We will still support the old slot syntax for backward compability,
by processing and converting the old slot syntax to the new
shorthand syntax.

Externally, users will be able to use the latest shorthand syntax for
slots as well, as our processing targets the latest shorthand syntax.