Code blocks: Support partial text highlighting (#1478)
The code blocks line highlighting functionality only supports full-text
and full-line highlights. Authors who wish to only highlight certain
words or key variables are not able to do so with the supported syntax.

Let's add a new highlight syntax to provide authors with the ability to
partially highlight text in a line.