Support printing of all tabs (#1498)
Tabs are supposed to contain alternative contents.
Readers can click on each tab to see its content. However,
when a page is printed or converted to PDF, the reader can
only see the content of the active tab.

Let's display all the tabs and its content when it is printed.