Compile Vue Page at build time to alleviate FOUC (#1512)
Compiling each Vue page in the browser may be expensive as it is
possible for a page to get very large in size. This creates the problem
of flash-of-unstyled-content (FOUC) for the user which is unsightly
and may affect user experience. Although there are temporary styles
in place to tackle the problem, the FOUC problem largely remains.

Let's compile every Vue page during build time to alleviate FOUC.