Implement live preview for pages in site.json (#1514)
When pages are added or removed from site.json, the respective
generated pages are not built or removed from the live preview.
Similarly, edited pages' attributes are not reflected as well. A server
reboot is necessary to display the new changes.

Let's support live preview for pages in site.json so that the user
does not have to restart the server each time page properties
are edited.