Change all $slots to $scopedSlots (#1535)
With the update to the deprecated slot syntax in #1464, the $slots
syntax does not work as it used to anymore, possibly due to the
internal changes for slot syntax in Vue. To be specific, named slots
will always disappear from $slots upon re-rendering but will always
be present in $scopedSlots. This causes named slot content to
disappear from Vue components upon any re-rendering.

Let's change all $slots syntax to $scopedSlots to prevent slot contents
from disappearing.