Implement live preview for site.json (#1529)
When attributes other than pages in site.json are modified, a server
restart is required to reflect the changes on the live preview. This 
can become a hassle to the user as they will need to frequently 
restart the server when trying out different site configuration.

Let's support live preview for all the attributes in site.json so 
that the latest changes can be reflected without restarting 
the server.