Add background build option to one-page serve (#1539)
Running markbind serve -o only builds pages as the user visits
those pages. As search data is populated according to built pages,
it is then limited to the pages that users have visited before,
rendering the search bar ineffective in live preview. Also, as
serve -o builds pages on-demand, the time between page builds is not
effectively utilized to build other pages.

Let's add an option to serve that can utilize the time after the
landing page has been generated to do a background build of pages that
are not yet built, updating the search data as it goes. Users thus get
not just the pages that they visited but also the pages that are built
in the background.