Lazy-serve: Rebuild opened pages on add/remove page (#1550)
With the addition of the multi-tab development flow, single-page serve
can keep track of all opened pages instead of a single currently
viewed page. This flow is already adopted in page edit events,
but it has yet to be adopted in page addition or removal.

Let's adopt the multi-tab development mechanism on the page addition' +
and removal flow.