Hoist <script> and <style> tags to after <body> (#1565)
Our setup compiles whatever is in the content section of the page into
Vue application. We also do server-side rendering on the content. Within
the content section, we allow users to write <script> and/or <style>
tags. However, during compilation, Vue will ignore and discard these tags.
Furthermore, during server-side rendering, special characters in
<script> tags will be escaped. While we have measures to tackle the
problems above, we should look to implement a better workaround.

Let's remove the temporary measures and prevent the compilation and
rendering of <script> and/or <style>, by hoisting them out of the
content section and to after <body> tag.