Handle baseUrl in live-preview reload check (#1594)
In live-preview, checks are implemented to only reload pages that
correspond to the changed file. However, the comparison used for the
check has not taken into account the base url of the site, which
causes these checks to always fail and thus pages are not automatically
reloaded.

Let's add base url handling to the check, so that sites with specified
can properly benefit from the live-preview feature.