Fix unresponsive fixed header padding (#1595)
When external stylesheets from different domains are used,
the cssRules of these stylesheets cannot be accessed or
modified due to CORS policy. This causes the method that
adjusts the fixed header padding to throw an error and
terminate prematurely.

Let's wrap this method in a try/catch block so that external
stylesheets from other domains will be skipped if cssRules
is not accessible.