Enhance appearance of minmised panels
* Add caret icon (as most minimised panels tend to use a caret icon to
show that it is a clickable button

* Add top margin (as expanded cards also have a top margin but not a
bottom margin, top margin was chosen. This is to prevent ugly cluttering)

Note: I think for code maintenance, we could probably remove some of the
styles in NestedPanel.vue . If I remember correctly, we only need to specify
the styles to change in the responsive breakpoints, so is it really needed
to specify the same style twice?

Note 2: https://developer.mozilla.org/en-US/docs/Web/CSS/float suggests that
it might be unneeded to specify display: inline-block if we are also
specifying float, as it would be overridden and ignored.