Update attribute to classes for consistency
The mix of attributes and style classes can be
confusing when users want to customize a hr element.

Let's fix the error of multiple classes on hr element
by updating the markdown-it-attrs package version and
change the syntax from `dotted` to `.dotted` etc.

This helps when users need to also customize the color
of the hr element, e.g. `--- {.dotted .border-primary}`.

The `markdown-it-attrs-nunjucks.js` is also updated
with the latest changes from `markdown-it-attrs`.