Forward markdown-it changes to patched files
The Markdown-it changelog mentions a few updates
which may affect some of the patches that we keep
in the packages/core/src/lib/markdown-it folder.

Let's update some of the content to match the
latest changes.

The markdown-it-attrs package has to be updated
together due to it's peer dependency requirment
on markdown-it. The update on the plugin is also
required to fix a reported bug.