Escape variables in PageNav from vue rendering
The raw variable interpolation syntax can be displayed
by wrapping it  with {% raw %} {% endraw %}. Even though
the {{ content }} is showing up correctly in the main page,
the generated PageNav text retains {{}} and causes vue
to incorrectly hide it due to rendering error.

Let's add a v-pre attribute whenever there is {{}} in
the navigation text.

This will allow content wrapped {{}} to display in PageNav.