Adjust line numbers setting for code blocks (#1734)
Line numbers are provided by default and there is no way
to override it globally.

Let's include a codeLineNumbers option in site.json to
enable a global configuration of line numbers. Line numbers
are also now hidden by default.

This is more convenient for authors as line numbers tend to
be hidden manually in practice. This is also consistent with other
markdown renderers.