Update outdated information in docs  (#1777)
Currently, the effect attribute is not implemented in popovers and 
tooltips, which means that using this attribute has no effect. 
However, the effect attribute is currently included in the examples 
for popovers. This may be confusing for readers. 

Let's remove all occurrences of the effect attribute from the docs 
in places where it has no effect. This will prevent readers from 
mistaking it as a valid attribute. 

Other minor grammatical errors were fixed as well.