Remove unused effect attribute from tests (#1783)
Currently, the effect attribute is not implemented for both popovers 
and tooltips. However, it is included in some of the tests for these 
components. This may be confusing and clutter the codebase. 

Let's remove the effect attribute from these tests, to avoid confusion 
on whether or not this is a valid attribute.