Use vue-final-modal for modal components (#1849)
Modals are implemented using bootstrap-vue directives.

bootstrap-vue has dependencies on specific versions of Bootstrap and
Vue, making it hard for MarkBind to migrate to Bootstrap 5, or Vue 3.

vue-final-modal has minimal dependencies, and supports both Vue 2 and 3.

Let's use vue-final-modal for modals instead. Doing so allows for future
migrations to Bootstrap 5 or Vue 3, without having to wait for
bootstrap-vue to migrate first. 

Vue is also upgraded from v2.6.11 to v2.6.14, to meet the dependencies
of vue-final-modal v2.4.1.