Report broken links only once (#1819)
When a link is validated and found to be broken, a warning is logged for
each time such a link is found. 

This makes the logs messy and harder to read while not providing
actionable information on which links to fix. Furthermore, logging each
link while the site is building will disrupt the CLI progress bar.

Let's use a site-wide abstraction to validate all the links at once and
report only unique broken links. Doing so allows us to avoid duplicate
link reporting and create a cleaner logging output.