Add Graphviz check when PlantUML diagrams are used (#1912)
Without Graphviz, PlantUML diagrams generated may
contain errors. Users may not be aware that diagrams are
generated incorrectly if they do not serve up the pages.

Let's add a check and warn users in the console if the
installation of Graphviz is missing.

This helps to keep users informed in the console when
the needed installation is missing. Users are also informed
of a way to disable the warning in `site.json`.