Refactor math formulae support (#1940)
The `dollars` delimiter for math formulae is currently supported via
a plugin because the markdown-it-texmath package does not
allow for more than one delimiter.

With the latest version update, the markdown-it-texmath package
can now support multiple delimiters.

Let's update the code to specify `dollars` as a supported delimiter
out of the box for math formulae. Let's also repurpose the
`texWithDollars` plugin to `mathDelimiters` for any additional
delimiter addition if any user is interested.

Users can now use `dollars` and `brackets` (two popular) delimiters
for math formulae without any configurations.