Discontinue the use of TypeScript project references (#1934)
The project references feature of TypeScript unfortunately comes with
quirks that hinder developer experience, such as not rebuilding
compiled files in certain cases.

Let's move away from the project references feature and opt for the
normal compilation instead. As a consequence of discontinuing the
feature, there is also a need for a script to clean compiled
files. Let's write the script and attach it to the post-checkout
hook to ease development.