Implement sticky header (#1939)
Since no assumptions of the user's header's height can be made at build
time, runtime style detection is required to calculate the header
height used in the fixed-header-padding class. This causes a flash of
unstyled content briefly before this detection is done
(the content is hidden behind the header).

Let's implement a sticky header instead, which "natively" offsets
content below it.

It should be noted that header height detection is still required
however, particularly for offsetting the sticky positioning of the
navigation menus on the side (top: var(--sticky-header-height)), since
they are (usually) much shorter than the main content element.

This header height is now exposed using the css variable
var(--sticky-header-height) for use in the user layout files.