Fix FOUC in navbar and search bar (#2004)
The placeholderDiv in Searchbar.vue is hidden using the
visibility property. The placeholderDiv will still take up space
on the navbar which will increase its height during loading.
Furthermore, the width of the search bar is set dynamically
during page load which results in FOUC in the search bar as
well.

Let's change the approach such that the placeholderDiv is hidden
and does not take up any height to prevent FOUC in the navbar.
The approach of setting the search bar width should also be
changed such that it is not set during page load to prevent
FOUC in the search bar.