Fix and update Algolia plugin (#2011)
On mobile view, if Algolia search is opened after the page
has been scrolled, the modal will appear at the top of the
page instead of the viewport. Users will be unable to close
the modal. In addition, Algolia has since released a stable
version of v3 docsearch.

Let's fixed the position of the search container so that the
modal always opens within the viewport. Let's also update
Algolia to the latest stable version.