Fix image linkify and alt attribute (#2046)
Fix image linkify and alt attribute

Currently, the image linkify feature is not documented and does
not account for already wrapped images. The image alternative
text is also not captured due to outdated markdown-it-imsize.

Let's document the linkify feature and update the relevant
packages to fix the issues. Let's also add tests to ensure the
functionalities are captured and verified.