Update embed macro to add omitFrontmatter (#2047)
In docs/userGuide/settingSiteProperties.md , the site title is set to
site.json File instead of Setting Site Properties. This is due to the
embed macro using <include /> without the omitFrontmatter
attribute.

Let's add an extra parameter to the embed macro to allow us to
include the omitFrontmatter attribute to <include /> if needed.