Deprecate TipBox (#2121)
Both tip-box and box are supported with box being an alias for
tipbox.

This could be confusing for users and developers as the use of
tip-box is not documented.

Let's deprecate tipbox, remove support for the tip-box tag and
replace all instances of tipbox to box.

This prevents confusion between the use of variations of tipbox in
both code and documentation, making any reference to or use of this
this component consistent. Replacing tipbox with box removes the
need for box being an alias for tipbox, deprecating tipbox completely.