Add <page-nav-print /> to allow printing (#2069)
Ability to print page-nav was removed as it disrupts the print
layout and was not useful as-is. However, printed pages do not
have access to page-nav and readers may still want an easy way
to navigate.

Let's add a <page-nav-print /> component to allow specifying
a position to print page-nav. This component is only visible on
print.

Having the page-nav component to be printed can be useful
as it serves as table of content. It needs to be added manually as
there may not be a good default position to place it on all pages.