Integrate TS classes in core package (#2176)
Integrate TS classes in core package

Most of the classes in the core package have been migrated to TS.
However, these classes are not integrated with each other as they are
not using each others' type definitions where applicable. 

Let's have the classes fully utilise other defined TS classes. Doing so
helps us fully reap the benefits of type safety and adds an extra layer
of checks and security to the code.