Refactor core package to use Node types (#2221)
Refactor core package to use Node types

In #2201, we introduced a unifying type for nodes. Most of the core
package is not yet using the new types.

Let's refactor the core package to utilise these new types, and add a
utility parseHTML method that avoids further `as unknown as` typecasts
throughout the codebase. Doing so allows us to make full use of
TypeScript's type-checking abilities while avoiding code duplication.

For more information on the rationale of the unifying types, refer to
#2201.