Upgrade lerna to version 6.5.1 (#2196)
Upgrade lerna to version 6.5.1

Our current version of lerna (version 4) is incompatible with
Node 19 due to an outdated dependency called node-gyp.
To ensure that we can support all versions of Node,
we need to upgrade to the latest version of lerna (version 6).

In addition to Node 19 compatibility,
upgrading to lerna 6 will allow us to take advantage of
new features like caching and concurrency.

During the upgrade process,
* I also noticed some formatting issues
caused by js-beautify v1.14.7 when using custom tag
(specifically custom tag with dash eg. site-nav). To fix this,
 I fixed the version of js-beautify to be v1.14.3 in package.json.
* I also realised there is some issue
with @types/node v17.0.22 so I updated it to v18.15.0.

To achieve these improvements, I made the following changes:
* Upgraded lerna 4 to lerna v6.5.1
* Fixed js-beautify version to v1.14.3
* Updated @types/node to v18.15.0
* Set up npm workspaces, init new package lock for workspaces 
and remove deprecated lerna bootsrap command
* Cached test results and run test/updatetest concurrently

This fixes the following issues:
* #1984: Explore using Turborepo to improve build system
* #2162: Node v19.6.0 incompatibility on Windows 10

These changes will enable users to
run our application on Node 19
and improve performance with caching and concurrency.