Adapt Site/index to TypeScript
Site/index.ts was previously bugged and caused errors with markbind
deploy. As a stopgap measure, Site/index.ts was re-adapted back to JS.

The bug has now been identified along with a fix, however,
Site/index will need to be remigrated back to TS. The false negative
test case in deploy has also been identifed and will be fixed as well.

Let's remigrate Site/index to TS, fix the deploy option that caused
the markbind deploy bug and also fix the negative test case that failed
to detect the bug.

Migration was made according to the original migration PR and the latest
v5.0.0.