Simplify fix for abrupt panel transition
The previous fix implemented a method to compute the max
height of a panel by checking if the collapse button is present
and adding the height of the bottom margin of the button.

Let's simplify this by wrapping the button in a <div> block and 
adding padding to the block. Let's also remove the getMaxHeight
method and its associated tests.