Remove Overridden Question Attributes in Docs (#2513)
The docs currently has a question where there are slots overriding
the header attributes. However, when #2511 is merged, logger
warnings about this overriding will be triggered, leading to the
check-docs GitHub action failing.

Let's remove the header attributes so that they will not trigger
the logger warning. Let's also update the sample code to reflect
this change.