Fixed Cannot read property 'isCE' of null Vue err
Error was due to multiple vue instances and package dependencies.
Shifted vue under root package.json, and set other vue
dependencies as peerDependencies.

This helps ensure that the version used is the same and
that the same instance is used.

Look at https://github.com/vuejs/core/issues/4344 for related
discussion on the topic.