Move IP validity check to earlier in the code
Currently, the validity check comes after the IP
zero check. The IP zero check flags some invalid
IPs, ,like `0`. This results in the following
scenario for a user:

1. User enters `0` as the IP
2. User gets a warning that the address `0.0.0.0`
has security risks
3. User proceeds by entering 'y'
4. User gets an error that the IP is invalid

This may be confusing for the user, who gets a
warning for their address, only to find out that
it's invalid. This commit moves the validity check
to earlier in the code, so that the user gets the
error message first and that the zero IP regex
check is not run on invalid IPs.