Add Tag colouring for card component
Tag colouring help to provide better visual appeal to the tags.

Due to the use of bootstrap badge for tags, the choice of color is
limited to 8 different colors. If there are more than 8 tags, the
color will recycle, making the tag coloration non unique.

In future, we can consider
1. adding more color schemes.
2. making the tags clickable as a button to filter search cards.