Fix page-nav, site-nav whitespace issue
This is due to differences in mechanism of portal vue
from v2 to v3, pertinently the removal of rendering of
a parent root component, causing the styling differences

To fix this regression, a parent div is added to simulate
the mechanism from v2. Further action can consider refactoring
this, however, the change is not too trivial and requires
understanding of current styling in main.css and account
for differences.