Improve error messages for `includePanelProcessor.ts` (#2703)
Currently, the error messages for includePanelProcessor.ts log both
the error message and the error stack. This may be confusing and
intimidating to users as it exposes some unnecessary details to them.

Let's update the logger to only log the error.message to hide the
error stack. Let's also update the error message for URLs in src to
be more user-friendly.